kobak_project/web-db/requirements.txt
streamlit==1.31.0
psycopg2-binary==2.9.9
pandas==1.5.0
extra_streamlit_components==0.1.60
streamlit-authenticator==0.3.1